Update CategoryRepository and CategoryDTO, and create CategoryService

+++ src/main/java/tc3005b224/amazonconnectinsights/repository/CategoryRepository.java
@@ -4,5 +4,5 @@ import org.springframework.stereotype.Repository;
 import tc3005b224.amazonconnectinsights.models_sql.Category;
 
 @Repository//
-public interface CategoryRepository extends CrudRepository<Category, Byte> {
+public interface CategoryRepository extends CrudRepository<Category, Short> {
 }
